Moving ZUGFeRD library to samples directory
