+++ Ghostscript.NET.PDFConverter.Samples/ZUGFeRD/Contact.cs
@@ -1,6 +1,6 @@
 using System.Xml;
 
-namespace Ghostscript.NET.FacturX.ZUGFeRD
+namespace Ghostscript.NET.PDFConverter.Samples.ZUGFeRD
 {
 	/// <summary>
 	///*
