+++ Ghostscript.NET.PDFConverter.Samples/ZUGFeRD/IZUGFeRDExportableContact.cs
@@ -1,7 +1,7 @@
-namespace Ghostscript.NET.FacturX.ZUGFeRD
+namespace Ghostscript.NET.PDFConverter.Samples.ZUGFeRD
 {
 
-	public interface IZUGFeRDExportableTradeParty
+	public interface IZUGFeRDExportableContact
 	{
 
 		/// <summary>
@@ -13,49 +13,42 @@ namespace Ghostscript.NET.FacturX.ZUGFeRD
 			return null;
 		}
 
-		/// <summary>
-		/// customer global identification assigned by the seller
-		/// </summary>
-		/// <returns> customer identification </returns>
-		string getGlobalID()
-		{
-			return null;
-		}
-
-		/// <summary>
-		///*
-		/// gets the official representation </summary>
-		/// <returns> the interface with the attributes of the legal organisation </returns>
-		/*	IZUGFeRDLegalOrganisation getLegalOrganisation()
-			{
-				return null;
-			}
-			*/
-		/// <summary>
-		/// customer global identification scheme
-		/// </summary>
-		/// <returns> customer identification </returns>
-		string getGlobalIDScheme()
-		{
-			return null;
-		}
-
-
-		IZUGFeRDExportableContact getContact()
-		{
-			return null;
-		}
 
 		/// <summary>
 		/// First and last name of the recipient
 		/// </summary>
 		/// <returns> First and last name of the recipient </returns>
-		string getName();
+		string? getName()
+		{
+			return null;
+		}
+
+		string getPhone()
+		{
+			return null;
+		}
+
+		string getEMail()
+		{
+			return null;
+		}
+
+		string getFax()
+		{
+			return null;
+		}
+
+
 		/// <summary>
 		/// Postal code of the recipient
 		/// </summary>
 		/// <returns> Postal code of the recipient </returns>
-		string getZIP();
+		string getZIP()
+		{
+			return null;
+		}
+
+
 		/// <summary>
 		/// VAT ID (Umsatzsteueridentifikationsnummer) of the contact
 		/// </summary>
@@ -64,21 +57,37 @@ namespace Ghostscript.NET.FacturX.ZUGFeRD
 		{
 			return null;
 		}
+
+
 		/// <summary>
 		/// two-letter country code of the contact
 		/// </summary>
 		/// <returns> two-letter iso country code of the contact </returns>
-		string getCountry();
+		string getCountry()
+		{
+			return null;
+		}
+
+
 		/// <summary>
 		/// Returns the city of the contact
 		/// </summary>
 		/// <returns> Returns the city of the recipient </returns>
-		string getLocation();
+		string getLocation()
+		{
+			return null;
+		}
+
+
 		/// <summary>
 		/// Returns the street address (street+number) of the contact
 		/// </summary>
 		/// <returns> street address (street+number) of the contact </returns>
-		string getStreet();
+		string getStreet()
+		{
+			return null;
+		}
+
 		/// <summary>
 		/// returns additional address information which is display in xml tag "LineTwo"
 		/// </summary>
@@ -87,13 +96,6 @@ namespace Ghostscript.NET.FacturX.ZUGFeRD
 		{
 			return null;
 		}
-		/// <summary>
-		///*
-		/// obligatory for sender but not for recipient </summary>
-		/// <returns> the tax id as string </returns>
-		string getTaxID()
-		{
-			return null;
-		}
+
 	}
 }
