+++ Ghostscript.NET.PDFConverter.Samples/ZUGFeRD/IZUGFeRDExportableItem.cs
@@ -1,5 +1,5 @@
 using System;
-namespace Ghostscript.NET.FacturX.ZUGFeRD
+namespace Ghostscript.NET.PDFConverter.Samples.ZUGFeRD
 {
 	public interface IZUGFeRDExportableItem
 	{
