+++ Ghostscript.NET.PDFConverter.Samples/ZUGFeRD/IZUGFeRDExportableTradeParty.cs
@@ -1,7 +1,7 @@
-namespace Ghostscript.NET.FacturX.ZUGFeRD
+namespace Ghostscript.NET.PDFConverter.Samples.ZUGFeRD
 {
 
-	public interface IZUGFeRDExportableContact
+	public interface IZUGFeRDExportableTradeParty
 	{
 
 		/// <summary>
@@ -13,42 +13,49 @@ namespace Ghostscript.NET.FacturX.ZUGFeRD
 			return null;
 		}
 
+		/// <summary>
+		/// customer global identification assigned by the seller
+		/// </summary>
+		/// <returns> customer identification </returns>
+		string getGlobalID()
+		{
+			return null;
+		}
+
+		/// <summary>
+		///*
+		/// gets the official representation </summary>
+		/// <returns> the interface with the attributes of the legal organisation </returns>
+		/*	IZUGFeRDLegalOrganisation getLegalOrganisation()
+			{
+				return null;
+			}
+			*/
+		/// <summary>
+		/// customer global identification scheme
+		/// </summary>
+		/// <returns> customer identification </returns>
+		string getGlobalIDScheme()
+		{
+			return null;
+		}
+
+
+		IZUGFeRDExportableContact getContact()
+		{
+			return null;
+		}
 
 		/// <summary>
 		/// First and last name of the recipient
 		/// </summary>
 		/// <returns> First and last name of the recipient </returns>
-		string? getName()
-		{
-			return null;
-		}
-
-		string getPhone()
-		{
-			return null;
-		}
-
-		string getEMail()
-		{
-			return null;
-		}
-
-		string getFax()
-		{
-			return null;
-		}
-
-
+		string getName();
 		/// <summary>
 		/// Postal code of the recipient
 		/// </summary>
 		/// <returns> Postal code of the recipient </returns>
-		string getZIP()
-		{
-			return null;
-		}
-
-
+		string getZIP();
 		/// <summary>
 		/// VAT ID (Umsatzsteueridentifikationsnummer) of the contact
 		/// </summary>
@@ -57,37 +64,21 @@ namespace Ghostscript.NET.FacturX.ZUGFeRD
 		{
 			return null;
 		}
-
-
 		/// <summary>
 		/// two-letter country code of the contact
 		/// </summary>
 		/// <returns> two-letter iso country code of the contact </returns>
-		string getCountry()
-		{
-			return null;
-		}
-
-
+		string getCountry();
 		/// <summary>
 		/// Returns the city of the contact
 		/// </summary>
 		/// <returns> Returns the city of the recipient </returns>
-		string getLocation()
-		{
-			return null;
-		}
-
-
+		string getLocation();
 		/// <summary>
 		/// Returns the street address (street+number) of the contact
 		/// </summary>
 		/// <returns> street address (street+number) of the contact </returns>
-		string getStreet()
-		{
-			return null;
-		}
-
+		string getStreet();
 		/// <summary>
 		/// returns additional address information which is display in xml tag "LineTwo"
 		/// </summary>
@@ -96,6 +87,13 @@ namespace Ghostscript.NET.FacturX.ZUGFeRD
 		{
 			return null;
 		}
-
+		/// <summary>
+		///*
+		/// obligatory for sender but not for recipient </summary>
+		/// <returns> the tax id as string </returns>
+		string getTaxID()
+		{
+			return null;
+		}
 	}
 }
