+++ Ghostscript.NET.PDFConverter.Samples/ZUGFeRD/Invoice.cs
@@ -21,7 +21,7 @@
  using System;
 using System.Collections.Generic;
 
-namespace Ghostscript.NET.FacturX.ZUGFeRD
+namespace Ghostscript.NET.PDFConverter.Samples.ZUGFeRD
 {
     /// <summary>
     ///*
