+++ Ghostscript.NET.PDFConverter.Samples/ZUGFeRD/LineCalculator.cs
@@ -1,5 +1,5 @@
 
-namespace Ghostscript.NET.FacturX.ZUGFeRD
+namespace Ghostscript.NET.PDFConverter.Samples.ZUGFeRD
 {
 
 	public class LineCalculator
