+++ Ghostscript.NET.PDFConverter.Samples/ZUGFeRD/TradeParty.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace Ghostscript.NET.FacturX.ZUGFeRD
+namespace Ghostscript.NET.PDFConverter.Samples.ZUGFeRD
 {
 
 	/// <summary>
