+++ Ghostscript.NET.PDFConverter.Samples/ZUGFeRD/VATAmount.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Ghostscript.NET.FacturX.ZUGFeRD
+namespace Ghostscript.NET.PDFConverter.Samples.ZUGFeRD
 {
 
 	public class VATAmount
